# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Api Book",
        "description": "desplay books whith gender and author",
        "contact": {
            "email": "adrien.charrier@outlook.fr"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "/genres": {
            "get": {
                "tags": [
                    "genre"
                ],
                "summary": "Get List Of Genres",
                "description": "Returns all Genres.",
                "operationId": "GetAllGenres",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "genre"
                ],
                "summary": "New genre",
                "description": "Put a new genre in the database.",
                "operationId": "Newgenre",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "name of the new genre",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/genres/{id}": {
            "get": {
                "tags": [
                    "genre"
                ],
                "summary": "Get A Genre",
                "description": "Get A Genre",
                "operationId": "GetAGenre",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Genre id",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "not found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "genre"
                ],
                "summary": "Update Genre",
                "description": "Update a genre in the database.",
                "operationId": "UpdateGenre",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Genre id",
                        "required": true
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "name",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/genre/{id}": {
            "delete": {
                "tags": [
                    "genre"
                ],
                "summary": "Delete A Genre",
                "description": "Delete A Genre",
                "operationId": "DeleteAGenre",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Book id",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/register": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "register",
                "description": "register",
                "operationId": "register",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "name",
                        "required": true
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "email",
                        "required": true
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "password",
                        "required": true
                    },
                    {
                        "name": "c_password",
                        "in": "query",
                        "description": "copy password",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            }
        },
        "/login": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "login",
                "description": "login",
                "operationId": "login",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "description": "email",
                        "required": true
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "password",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "Bearer"
            }
        }
    },
    "security": [
        []
    ]
}